*hannes: syntax
git-svn-id: file:///usr/local/Singular/svn/trunk@9085 2c84dea3-7e68-4137-9b89-c4e89433aadc

+++ kernel/misc.cc
@@ -74,7 +74,7 @@ omBin ip_sring_bin = omGetSpecBin(sizeof(ip_sring));
 * the global exit routine of Singular
 */
 extern "C" {
-void m2_end(short i)
+void m2_end(int i)
 {
   fe_reset_input_mode();
   #ifdef PAGE_TEST
